# Replace regex validators by a faster equivalent implementation.