--- a/conjure-core/src/main/java/com/palantir/conjure/defs/validator/PackageValidator.java
+++ b/conjure-core/src/main/java/com/palantir/conjure/defs/validator/PackageValidator.java
@@ -20,9 +20,9 @@ import com.google.common.base.Joiner;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableList;
-import com.palantir.conjure.CachedPatternMatcher;
+import com.palantir.conjure.PackagePattern;
+import com.palantir.conjure.SimplifiedPattern;
 import java.util.List;
-import java.util.regex.Pattern;
 
 public final class PackageValidator {
 
@@ -31,16 +31,13 @@ public final class PackageValidator {
 
     private PackageValidator() {}
 
-    private static final CachedPatternMatcher VALID_PACKAGE =
-            CachedPatternMatcher.wrap(Pattern.compile("^([a-z][a-z0-9]+(\\" + ".[a-z][a-z0-9]*)*)?$"));
+    private static final SimplifiedPattern VALID_PACKAGE = new PackagePattern();
 
     public static List<String> components(String name) {
         return ImmutableList.copyOf(Splitter.on('.').split(name));
     }
 
     public static void validate(String name) {
-        // TODO(rfink): NPE when the pattern is static !?!?
-
         Preconditions.checkArgument(
                 VALID_PACKAGE.matches(name),
                 "Conjure package names must match pattern %s: %s",