… changelog/@unreleased/pr-999.v2.yml
… conjure-core/src/test/java/com/palantir/conjure/defs/validator/ParamIdValidatorTest.java
… conjure-generator-common/src/main/java/com/palantir/conjure/CharUtils.java
… conjure-generator-common/src/main/java/com/palantir/conjure/EnumPattern.java
… conjure-generator-common/src/main/java/com/palantir/conjure/KebabCasePattern.java
… conjure-generator-common/src/main/java/com/palantir/conjure/PackagePattern.java
… conjure-generator-common/src/main/java/com/palantir/conjure/SimplifiedPattern.java
… conjure-generator-common/src/main/java/com/palantir/conjure/SnakeCasePattern.java
… conjure-generator-common/src/test/java/com/palantir/conjure/CamelCasePatternTest.java
… conjure-generator-common/src/test/java/com/palantir/conjure/EnumPatternTest.java
… conjure-generator-common/src/test/java/com/palantir/conjure/KebabCasePatternTest.java
… conjure-generator-common/src/test/java/com/palantir/conjure/PackagePatternTest.java
… conjure-generator-common/src/test/java/com/palantir/conjure/PatternTestUtil.java
… conjure-generator-common/src/test/java/com/palantir/conjure/SnakeCasePatternTest.java
… versions.lock
… versions.props