## Before this PR
<!-- What's wrong with the current state of the world and why change it now? -->

Compiling large amounts of conjure takes a lot of time; for example, here's a sample run of one internal tool:

```
./gradlew $project:compileIr --rerun-tasks  1.43s user 0.23s system 2% cpu 1:09.24 total
```
I.e. 69s total

Main contributors from that sample run:
```
:$project:rawIr | 21.871s |  
:$project:compileIr | 21.492s |  
```

## After this PR
<!-- User-facing outcomes this PR delivers go below -->
==COMMIT_MSG==
Replace regex validators by a faster equivalent implementation.
==COMMIT_MSG==

```
./gradlew $project:compileIr --rerun-tasks  1.32s user 0.20s system 4% cpu 34.243 total
```
I.e. 34s total

Main contributors from that sample run:
```
:$project:rawIr | 9.597s |  
:$project:compileIr | 8.637s |
```

## Possible downsides?
<!-- Please describe any way users could be negatively affected by this PR. -->