@carterkozak 2021-12-01
I tend to prefer unrolling regex into java code, that would avoid the cache and likely be just as fast.