@raiju 2021-12-09
To validate your suspicions, coded up the camel case matcher manually & ran some JMH benchmarks:
```
Benchmark                     Mode  Cnt   Score   Error   Units
RegexBenchmark.manual        thrpt    5  44.126 ± 0.614  ops/us
RegexBenchmark.matcherReuse  thrpt    5   1.286 ± 0.035  ops/us
RegexBenchmark.regex         thrpt    5   1.201 ± 0.005  ops/us
```

Looks like we can hit a > 40x improvement by coding it manually, which should help a lot.