@raiju 2021-12-09
@carterkozak Changed it to hand-written implementations