@carterkozak 2021-12-09
> Looks like we can hit a > 40x improvement by coding it manually, which should help a lot.

Looks great, thanks for working on this and validating with benchmarks :-)