@carterkozak 2022-01-18
generally lgtm, just a couple of nits around redundant accessors