@svc-autorelease 2022-01-19
Released [4.22.0](https://github.com/palantir/conjure/releases/tag/4.22.0)